coconowbackend/requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgi-redis==1.4.3
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
autobahn==22.3.2
Automat==0.8.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.6
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==19.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
dbus-python==1.2.16
Deprecated==1.2.13
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-cors-headers==3.11.0
django-redis==5.2.0
django-rename-app==0.1.3
djangorestframework==3.13.1
djlint==1.0.1
entrypoints==0.3
environ==1.0
h11==0.13.0
hiredis==2.0.0
httpie==1.0.3
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.4.0
incremental==16.10.1
Jinja2==2.11.3
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.4.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
msgpack-python==0.5.6
netifaces==0.10.4
oauthlib==3.1.0
outcome==1.1.0
packaging==21.3
paramiko==2.10.4
pathspec==0.9.0
pexpect==4.6.0
Pillow==9.1.0
prompt-toolkit==3.0.29
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
pytz==2022.1
PyYAML==6.0
redis==4.3.1
regex==2022.4.24
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.5
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.2
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
tomli==2.0.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
txaio==22.2.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
zipp==3.7.0
zope.interface==4.7.1
